GEP 10: test each node at the seam the currency crossing now sits on
`processed_data` tests cover sorting, id remapping and dtype coercion and
take already-converted magnitudes. The conversion tests move to
`test_currency.py` and drive `input_data_in_computation_currency` directly,
plus direct coverage of `CurrencyConversion`'s identity case and its
object-dtype passthrough.
