+++ tests/interface_dag_elements/test_processed_data.py
@@ -1,30 +1,14 @@
 from __future__ import annotations
 
-import dags.tree as dt
 import numpy
 import pandas as pd
 import pytest
-from mettsim.middle_earth import UNIT_SYSTEM
 
 from ttsim.interface_dag_elements.input_data import sort_indices
 from ttsim.interface_dag_elements.processed_data import (
     _canonicalize_input_dtype,
     processed_data,
 )
-from ttsim.interface_dag_elements.specialized_environment import (
-    _add_derived_functions,
-)
-from ttsim.tt import TTSIMUnit, policy_input
-
-
-def _spec_env(policy_environment, input_data__flat, grouping_levels=()):
-    """The specialized environment the currency conversion reads units off."""
-    return _add_derived_functions(
-        qname_env_without_tree_logic=policy_environment,
-        tt_targets={},
-        input_columns={dt.qname_from_tree_path(p) for p in input_data__flat},
-        grouping_levels=grouping_levels,
-    )
 
 
 @pytest.fixture
@@ -45,15 +29,11 @@ def test_processed_data(input_data__flat, xnp):
     pd.testing.assert_frame_equal(
         pd.DataFrame(
             processed_data(
-                input_data__flat=input_data__flat,
+                input_data_in_computation_currency=input_data__flat,
                 input_data__sort_indices=sort_indices(
                     input_data__flat=input_data__flat, xnp=xnp
                 ),
                 xnp=xnp,
-                specialized_environment__without_tree_logic_and_with_derived_functions={},
-                data_currency="CASTAR",
-                computation_currency="CASTAR",
-                unit_system=UNIT_SYSTEM,
             )
         ),
         pd.DataFrame(expected),
@@ -78,13 +58,9 @@ def test_processed_data_foreign_key_out_of_bounds(xnp):
     pd.testing.assert_frame_equal(
         pd.DataFrame(
             processed_data(
-                input_data__flat=input_data__flat,
+                input_data_in_computation_currency=input_data__flat,
                 input_data__sort_indices=input_data__sort_indices,
                 xnp=xnp,
-                specialized_environment__without_tree_logic_and_with_derived_functions={},
-                data_currency="CASTAR",
-                computation_currency="CASTAR",
-                unit_system=UNIT_SYSTEM,
             )
         ),
         pd.DataFrame(expected),
@@ -109,13 +85,9 @@ def test_processed_data_foreign_key_inside_bounds(xnp):
     pd.testing.assert_frame_equal(
         pd.DataFrame(
             processed_data(
-                input_data__flat=input_data__flat,
+                input_data_in_computation_currency=input_data__flat,
                 input_data__sort_indices=input_data__sort_indices,
                 xnp=xnp,
-                specialized_environment__without_tree_logic_and_with_derived_functions={},
-                data_currency="CASTAR",
-                computation_currency="CASTAR",
-                unit_system=UNIT_SYSTEM,
             )
         ),
         pd.DataFrame(expected),
@@ -136,13 +108,9 @@ def test_processed_data_single_column(xnp):
     pd.testing.assert_frame_equal(
         pd.DataFrame(
             processed_data(
-                input_data__flat=input_data__flat,
+                input_data_in_computation_currency=input_data__flat,
                 input_data__sort_indices=input_data__sort_indices,
                 xnp=xnp,
-                specialized_environment__without_tree_logic_and_with_derived_functions={},
-                data_currency="CASTAR",
-                computation_currency="CASTAR",
-                unit_system=UNIT_SYSTEM,
             )
         ),
         pd.DataFrame(expected),
@@ -202,15 +170,11 @@ def test_processed_data_coerces_uint_columns_to_signed(xnp):
         ("wage",): numpy.array([0, 100], dtype=numpy.uint32),
     }
     result = processed_data(
-        input_data__flat=input_data__flat,
+        input_data_in_computation_currency=input_data__flat,
         input_data__sort_indices=sort_indices(
             input_data__flat=input_data__flat, xnp=xnp
         ),
         xnp=xnp,
-        specialized_environment__without_tree_logic_and_with_derived_functions={},
-        data_currency="CASTAR",
-        computation_currency="CASTAR",
-        unit_system=UNIT_SYSTEM,
     )
     assert result["wage"].dtype.kind == "i"
     # Subtraction stays signed instead of underflowing into uint wraparound.
@@ -235,193 +199,10 @@ def test_processed_data_single_row(xnp):
     pd.testing.assert_frame_equal(
         pd.DataFrame(
             processed_data(
-                input_data__flat=input_data__flat,
+                input_data_in_computation_currency=input_data__flat,
                 input_data__sort_indices=input_data__sort_indices,
                 xnp=xnp,
-                specialized_environment__without_tree_logic_and_with_derived_functions={},
-                data_currency="CASTAR",
-                computation_currency="CASTAR",
-                unit_system=UNIT_SYSTEM,
             )
         ),
         pd.DataFrame(expected),
     )
-
-
-def test_processed_data_converts_pint_tagged_currency_input(xnp):
-    """Layer-2 boundary (GEP 10): a tagged input is converted to the data currency.
-
-    End-to-end through the ``processed_data`` interface node: a wealth column
-    handed in as silver pennies rides along castar data and is rescaled at the
-    boundary (4 silver pennies = 1 castar).
-    """
-    input_data__flat = {
-        ("p_id",): numpy.array([0, 1]),
-        ("wealth",): UNIT_SYSTEM.registry.Quantity(
-            numpy.array([4.0, 8.0]), "SILVER_PENNY"
-        ),
-    }
-    out = processed_data(
-        input_data__flat=input_data__flat,
-        input_data__sort_indices=sort_indices(
-            input_data__flat=input_data__flat, xnp=xnp
-        ),
-        xnp=xnp,
-        specialized_environment__without_tree_logic_and_with_derived_functions={},
-        data_currency="CASTAR",
-        computation_currency="CASTAR",
-        unit_system=UNIT_SYSTEM,
-    )
-    assert not isinstance(out["wealth"], UNIT_SYSTEM.registry.Quantity)
-    assert list(numpy.asarray(out["wealth"])) == pytest.approx([1.0, 2.0])
-
-
-def test_processed_data_converts_untagged_currency_input_by_declared_unit(xnp):
-    """Untagged input data converts, too (GEP 10).
-
-    Data arrives in the data currency; every column whose *declared* unit
-    carries a currency component is converted to the computation currency.
-    Non-currency columns are untouched.
-    """
-
-    @policy_input(unit=TTSIMUnit.CURRENCY)
-    def wealth() -> float:
-        pass
-
-    @policy_input(unit=TTSIMUnit.YEARS)
-    def age() -> int:
-        pass
-
-    input_data__flat = {
-        ("p_id",): numpy.array([0, 1]),
-        ("wealth",): numpy.array([4.0, 8.0]),
-        ("age",): numpy.array([30, 40]),
-    }
-    out = processed_data(
-        input_data__flat=input_data__flat,
-        input_data__sort_indices=sort_indices(
-            input_data__flat=input_data__flat, xnp=xnp
-        ),
-        xnp=xnp,
-        specialized_environment__without_tree_logic_and_with_derived_functions=(
-            _spec_env(
-                policy_environment={"wealth": wealth, "age": age},
-                input_data__flat=input_data__flat,
-                grouping_levels=("hh",),
-            )
-        ),
-        data_currency="SILVER_PENNY",
-        computation_currency="CASTAR",
-        unit_system=UNIT_SYSTEM,
-    )
-    assert list(numpy.asarray(out["wealth"])) == pytest.approx([1.0, 2.0])
-    assert list(numpy.asarray(out["age"])) == [30, 40]
-
-
-def test_derived_input_ignores_siblings_a_derivation_cannot_produce(xnp):
-    """`income_hh` aggregates the declared non-currency `income`; the currency
-    flow `income_m` shares the base name but no derivation produces `income_hh`
-    from it (a time suffix is only ever rebased, never dropped), so the column
-    must not convert."""
-
-    @policy_input(unit=TTSIMUnit.CURRENCY.PER_MONTH)
-    def income_m() -> float:
-        pass
-
-    @policy_input(unit=TTSIMUnit.YEARS)
-    def income() -> int:
-        pass
-
-    input_data__flat = {
-        ("p_id",): numpy.array([0, 1]),
-        ("income_hh",): numpy.array([4.0, 4.0]),
-    }
-    out = processed_data(
-        input_data__flat=input_data__flat,
-        input_data__sort_indices=sort_indices(
-            input_data__flat=input_data__flat, xnp=xnp
-        ),
-        xnp=xnp,
-        specialized_environment__without_tree_logic_and_with_derived_functions=(
-            _spec_env(
-                policy_environment={"income_m": income_m, "income": income},
-                input_data__flat=input_data__flat,
-                grouping_levels=("hh",),
-            )
-        ),
-        data_currency="SILVER_PENNY",
-        computation_currency="CASTAR",
-        unit_system=UNIT_SYSTEM,
-    )
-    assert list(numpy.asarray(out["income_hh"])) == pytest.approx([4.0, 4.0])
-
-
-def test_processed_data_converts_derived_input_via_its_stub(xnp):
-    """A derived name provided as data converts via its `PolicyInput` stub.
-
-    ``wage_m_hh`` has no declaration of its own; the stub minted from the
-    declared ``wage_m`` carries ``CURRENCY_PER_MONTH_PER_HH`` (aggregation
-    never adds or removes the currency component), so the column converts.
-    """
-
-    @policy_input(unit=TTSIMUnit.CURRENCY.PER_MONTH)
-    def wage_m() -> float:
-        pass
-
-    input_data__flat = {
-        ("p_id",): numpy.array([0, 1]),
-        ("wage_m_hh",): numpy.array([4.0, 4.0]),
-    }
-    out = processed_data(
-        input_data__flat=input_data__flat,
-        input_data__sort_indices=sort_indices(
-            input_data__flat=input_data__flat, xnp=xnp
-        ),
-        xnp=xnp,
-        specialized_environment__without_tree_logic_and_with_derived_functions=(
-            _spec_env(
-                policy_environment={"wage_m": wage_m},
-                input_data__flat=input_data__flat,
-                grouping_levels=("hh",),
-            )
-        ),
-        data_currency="SILVER_PENNY",
-        computation_currency="CASTAR",
-        unit_system=UNIT_SYSTEM,
-    )
-    assert list(numpy.asarray(out["wage_m_hh"])) == pytest.approx([1.0, 1.0])
-
-
-def test_processed_data_converts_time_variant_input_via_its_stub(xnp):
-    """Data at a time-converted name converts via its `PolicyInput` stub.
-
-    ``wage_y`` has no declaration of its own; the stub minted from the
-    declared ``wage_m`` carries ``CURRENCY_PER_YEAR``, so the column converts.
-    """
-
-    @policy_input(unit=TTSIMUnit.CURRENCY.PER_MONTH)
-    def wage_m() -> float:
-        pass
-
-    input_data__flat = {
-        ("p_id",): numpy.array([0, 1]),
-        ("wage_y",): numpy.array([4.0, 4.0]),
-    }
-    out = processed_data(
-        input_data__flat=input_data__flat,
-        input_data__sort_indices=sort_indices(
-            input_data__flat=input_data__flat, xnp=xnp
-        ),
-        xnp=xnp,
-        specialized_environment__without_tree_logic_and_with_derived_functions=(
-            _spec_env(
-                policy_environment={"wage_m": wage_m},
-                input_data__flat=input_data__flat,
-                grouping_levels=("hh",),
-            )
-        ),
-        data_currency="SILVER_PENNY",
-        computation_currency="CASTAR",
-        unit_system=UNIT_SYSTEM,
-    )
-    assert list(numpy.asarray(out["wage_y"])) == pytest.approx([1.0, 1.0])
